+++ frontend/components/FileTable/FileTable.tsx
@@ -8,119 +8,6 @@ import styled from "styled-components";
 import { File } from "../../lib/generated/graphql";
 import { FileIcon } from "../Icon";
 
-interface Props {
-  file: Pick<File, "title" | "id" | "fileType" | "folder" | "modifiedAt">;
-  workspaceId: string;
-  titleLink: boolean;
-}
-
-// Mobile
-const ListItem = styled.li`
-  align-items: flex-start;
-  display: flex;
-  padding-top: 16px;
-  padding-bottom: 16px;
-  border-bottom: 1px solid ${({ theme }) => theme.colorNhsukGrey4};
-`;
-
-const RHContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding-left: 12px;
-  > div {
-    padding-bottom: 20px;
-  }
-  p,
-  h4 {
-    margin-bottom: 0;
-    font-size: 18px;
-  }
-  a {
-    text-decoration: underline;
-    font-size: 16px;
-  }
-`;
-
-const MobileTitle = styled.p`
-  padding-bottom: 20px;
-  font-weight: 700;
-  color: ${({ theme }) => theme.colorNhsukBlack};
-`;
-
-const FileListItem = ({ file, workspaceId, titleLink }: Props) => {
-  const modifiedAt = format(parseISO(file.modifiedAt), "LLL d, yyyy");
-  return (
-    <ListItem>
-      <FileIcon fileType={file.fileType} />
-      <RHContainer>
-        {titleLink ? (
-          <MobileTitle>
-            <Link
-              href={`/workspaces/${workspaceId}/folders/${file.folder}/files/${file.id}`}
-              passHref
-            >
-              <a>{file.title}</a>
-            </Link>
-          </MobileTitle>
-        ) : (
-          <MobileTitle>{file.title}</MobileTitle>
-        )}
-
-        <div>
-          <h4>Last modified</h4>
-          <p>{modifiedAt}</p>
-        </div>
-        <a>Download file</a>
-      </RHContainer>
-    </ListItem>
-  );
-};
-
-const List = styled.ul`
-  padding-left: 12px;
-  padding-right: 16px;
-  ${({ theme }) => `
-    background: ${theme.colorNhsukGrey5};
-    @media (min-width: ${theme.mqBreakpoints.tablet}) {
-      display: none;
-    }
-  `}
-`;
-
-const Heading = styled.h3`
-  font-size: 1.5rem;
-  line-height: 1.33333;
-  background-color: #005eb8;
-  color: #ffffff;
-  display: inline-block;
-  margin: 0 0 8px -32px;
-  margin-top: 0px;
-  padding: 8px 32px;
-  position: relative;
-  top: -16px;
-`;
-
-export const MobileFileList = ({
-  files,
-  workspaceId,
-  titleLink,
-  tableHeading,
-}: FileTableProps) => (
-  <>
-    <Heading>{tableHeading}</Heading>
-    <List>
-      {files.map((file) => (
-        <FileListItem
-          key={file.id}
-          file={file}
-          workspaceId={workspaceId}
-          titleLink={titleLink}
-        />
-      ))}
-    </List>
-  </>
-);
-
 // Tablet and Desktop
 interface FileTableProps {
   files: Pick<File, "title" | "id" | "folder" | "fileType" | "modifiedAt">[];
@@ -138,7 +25,7 @@ const TableContainer = styled.div`
     }
   `}
   a {
-    text-decoration: underline;
+    /* text-decoration: underline; */
   }
   > div {
     ${({ theme }) => `
@@ -147,14 +34,14 @@ const TableContainer = styled.div`
   }
 `;
 
-const Title = styled.span`
-  font-weight: 700;
-`;
-
 const NHSTable = styled(Table)`
   tbody tr:hover {
     background: white;
   }
+`;
+
+const ModifiedDate = styled(Table.Cell)`
+  color: ${({ theme }) => theme.colorNhsukGrey1};
 `;
 
 export const FileTable = ({
@@ -189,18 +76,20 @@ export const FileTable = ({
                       passHref
                     >
                       <a>
-                        <Title>{file.title}</Title>
+                        <span>{file.title}</span>
                       </a>
                     </Link>
                   ) : (
-                    <Title>{file.title}</Title>
+                    <span>{file.title}</span>
                   )}
                 </Table.Cell>
-                <Table.Cell>{modifiedAt}</Table.Cell>
+                <ModifiedDate>{modifiedAt}</ModifiedDate>
                 <Table.Cell>
-                  <a style={{ display: "inline-block", paddingRight: "8px" }}>
-                    Download file
-                  </a>
+                  <Link href="" passHref>
+                    <a style={{ display: "inline-block", paddingRight: "8px" }}>
+                      Download file
+                    </a>
+                  </Link>
                 </Table.Cell>
               </Table.Row>
             );
