+++ frontend/components/FileTable/index.tsx
@@ -1,5 +1,6 @@
 import { File } from "../../lib/generated/graphql";
-import { MobileFileList, FileTable } from "./FileTable";
+import { FileTable } from "./FileTable";
+import { MobileFileList } from "./MobileFileTable";
 
 export { MobileFileList, FileTable };
 export type { File };
